Renamed SurveyFormView into SurveyForm
